Jabel, Generic injection and mostly automatic code cleanup (#1829)
* Enable Jabel&Generic injection, fix type error caused by this

* add missing <>

* Infer generic types automatically

* Parametrize cast types

* Use enhanced for loops

* Unnecessary boxing

* Unnecessary unboxing

* Use Objects.equals

* Explicit type can be replaced with `<>`

* Collapse identical catch blocks

* Add SafeVarargs where applicable

* Anonymous type can be replaced with lambda

* Use List.sort directly

* Lambda can be a method reference

* Statement lambda can be an expression lambda

* Use string switches

* Instanceof pattern matching

* Text block can be used

* Migrate to enhanced switch

* Java style array declarations

* Unnecessary toString()

* More unnecessary String conversions

* Unnecessary modifiers

* Unnecessary semicolons

* Fix duplicate conditions

* Extract common code from if branches

* Replace switches with ifs for 1-2 cases

* Inner class may be static

* Minor performance issues

* Replace string appending in loops with string builders

* Fix IntelliJ using the wrong empty list method

* Use Long.compare

* Generic arguments: getSubItems

* Generic arguments: getSubBlocks

* Raw types warnings

* Fix remaining missing generics

* Too weak variable type leads to unnecessary cast

* Redundant type casts

* Redundant array length check

* Redundant vararg arrays

* Manual min/max implementations

* A couple missed inspections

* Goodbye explosion power ternary ladder

* Apply spotless

* Get rid of the other two big ternary ladders

* Binary search explosion power

* Don't overcomplicate things

+++ src/main/java/gregtech/api/metatileentity/implementations/GT_MetaTileEntity_Transformer.java
@@ -179,7 +179,7 @@ public class GT_MetaTileEntity_Transformer extends GT_MetaTileEntity_TieredMachi
                             ForgeDirection.getOrientation(GT_Utility.getOppositeSide(i)),
                             1,
                             true) == 1) {
-                        long tEU = (long) ((IEnergyProvider) tTileEntity).extractEnergy(
+                        long tEU = ((IEnergyProvider) tTileEntity).extractEnergy(
                                 ForgeDirection.getOrientation(GT_Utility.getOppositeSide(i)),
                                 GT_Utility.safeInt(maxEUInput() * 100L / GregTech_API.mRFtoEU),
                                 false);
@@ -187,7 +187,7 @@ public class GT_MetaTileEntity_Transformer extends GT_MetaTileEntity_TieredMachi
                         aBaseMetaTileEntity.injectEnergyUnits((byte) 6, Math.min(tEU, maxEUInput()), 1);
                     } else if (tTileEntity instanceof IEnergyStorage
                             && ((IEnergyStorage) tTileEntity).extractEnergy(1, true) == 1) {
-                                long tEU = (long) ((IEnergyStorage) tTileEntity).extractEnergy(
+                                long tEU = ((IEnergyStorage) tTileEntity).extractEnergy(
                                         GT_Utility.safeInt(maxEUInput() * 100L / GregTech_API.mRFtoEU),
                                         false);
                                 tEU = tEU * GregTech_API.mRFtoEU / 100;
